+++ code/src/sixsq/nuvla/server/resources/module.clj
@@ -9,7 +9,6 @@ component, or application.
     [sixsq.nuvla.auth.utils :as auth]
     [sixsq.nuvla.db.filter.parser :as parser]
     [sixsq.nuvla.db.impl :as db]
-    [sixsq.nuvla.pricing.impl :as pricing-impl]
     [sixsq.nuvla.server.resources.common.crud :as crud]
     [sixsq.nuvla.server.resources.common.std-crud :as std-crud]
     [sixsq.nuvla.server.resources.common.utils :as u]
@@ -115,46 +114,6 @@ component, or application.
         crud/validate)
     {}))
 
-
-(defn active-claim->account-id
-  [active-claim]
-  (let [filter     (format "parent='%s'" active-claim)
-        options    {:cimi-params {:filter (parser/parse-cimi-filter filter)}}
-        account-id (-> (crud/query-as-admin "vendor" options)
-                       second
-                       first
-                       :account-id)]
-    (or account-id
-        (throw (r/ex-response (str "unable to resolve vendor account-id for active-claim '"
-                                   active-claim "' ") 409)))))
-
-
-(defn set-price
-  [{{:keys [price-id cent-amount-daily currency follow-customer-trial] :as price}
-    :price name :name path :path :as body}
-   active-claim]
-  (if price
-    (let [product-id (some-> price-id pricing-impl/retrieve-price
-                             pricing-impl/price->map :product-id)
-          account-id (active-claim->account-id active-claim)
-          s-price    (pricing-impl/create-price
-                       (cond-> {"currency"    currency
-                                "unit_amount" cent-amount-daily
-                                "recurring"   {"interval"        "month"
-                                               "aggregate_usage" "sum"
-                                               "usage_type"      "metered"}}
-                               product-id (assoc "product" product-id)
-                               (nil? product-id) (assoc "product_data" {"name"       (or name path)
-                                                                        "unit_label" "day"})))
-          price      (cond-> {:price-id          (pricing-impl/get-id s-price)
-                              :product-id        (pricing-impl/get-product s-price)
-                              :account-id        account-id
-                              :cent-amount-daily cent-amount-daily
-                              :currency          currency}
-                             (some? follow-customer-trial) (assoc :follow-customer-trial follow-customer-trial))]
-      (assoc body :price price))
-    body))
-
 (defn throw-cannot-access-registries-or-creds
   [{{{:keys [private-registries registries-credentials]} :content} :body :as request}]
   (when
@@ -189,35 +148,31 @@ component, or application.
              (count creds))
       (throw (r/ex-response "Registries credentials can't be resolved!" 403)))))
 
+(defn throw-compatibility-required-for-application
+  [{:keys [subtype compatibility]}]
+  (when (and (utils/is-application? subtype) (nil? compatibility))
+    (throw (r/ex-response "Application subtype should have compatibility attribute set!" 400))))
+
 
 (defmethod crud/add resource-type
   [{:keys [body] :as request}]
 
   (a/throw-cannot-add collection-acl request)
-
   (throw-colliding-path (:path body))
-
   (throw-cannot-access-registries-or-creds request)
-
   (throw-price-error body)
+  (throw-compatibility-required-for-application body)
 
   (let [[{:keys [subtype] :as module-meta}
-         {:keys [author commit docker-compose] :as module-content}] (-> body u/strip-service-attrs
-                                                                        utils/split-resource)
-        module-meta (dissoc module-meta :compatibility :parent-path :published)]
+         {:keys [author commit] :as module-content}] (-> body u/strip-service-attrs
+                                                         utils/split-resource)
+        module-meta (dissoc module-meta :parent-path :published)]
 
     (if (utils/is-project? subtype)
       (db-add-module-meta module-meta request)
       (let [content-url     (subtype->resource-url subtype)
 
-            [compatibility
-             unsupported-options] (utils/parse-get-compatibility-fields subtype docker-compose)
-
-            content-body    (-> module-content
-                                (dissoc :unsupported-options)
-                                (merge {:resource-type content-url})
-                                (cond-> (seq unsupported-options) (assoc :unsupported-options
-                                                                         unsupported-options)))
+            content-body    (merge module-content {:resource-type content-url})
 
             content-request {:params      {:resource-name content-url}
                              :body        content-body
@@ -230,12 +185,8 @@ component, or application.
             (assoc :versions [(cond-> {:href   content-id
                                        :author author}
                                       commit (assoc :commit commit))])
-            (set-price (auth/current-active-claim request))
-            (cond-> compatibility (assoc :compatibility compatibility))
+            (utils/set-price (auth/current-active-claim request))
             (db-add-module-meta request))))))
-
-
-
 
 
 (defn retrieve-edn
@@ -252,6 +203,7 @@ component, or application.
           is-not-project? (not (utils/is-project? subtype))]
       (-> module-meta
           (cond-> is-not-project? (utils/get-module-content uuid))
+          utils/resolve-vendor-email
           (crud/set-operations request)
           (a/select-viewable-keys request)
           (r/json-response)))
@@ -283,12 +235,12 @@ component, or application.
   (try
     (let [id          (str resource-type "/" (-> request :params :uuid))
           [module-meta
-           {:keys [author commit docker-compose] :as module-content}] (-> body
-                                                                          u/strip-service-attrs
-                                                                          utils/split-resource)
+           {:keys [author commit] :as module-content}] (-> body
+                                                           u/strip-service-attrs
+                                                           utils/split-resource)
           {:keys [subtype versions price acl]} (crud/retrieve-by-id-as-admin id)
           module-meta (-> module-meta
-                          (dissoc :compatibility :parent-path :published)
+                          (dissoc :parent-path :published)
                           (assoc :subtype subtype)
                           utils/set-parent-path)]
 
@@ -301,16 +253,7 @@ component, or application.
         (let [_              (throw-cannot-access-registries-or-creds request)
               content-url    (subtype->resource-url subtype)
 
-              [compatibility
-               unsupported-options] (when docker-compose
-                                      (utils/parse-get-compatibility-fields
-                                        subtype docker-compose))
-
-              content-body   (some-> module-content
-                                     (dissoc :unsupported-options)
-                                     (merge {:resource-type content-url})
-                                     (cond-> (seq unsupported-options) (assoc :unsupported-options
-                                                                              unsupported-options)))
+              content-body   (some-> module-content (merge {:resource-type content-url}))
 
               content-id     (when content-body
                                (-> {:params      {:resource-name content-url}
@@ -330,14 +273,14 @@ component, or application.
                                             (get-in module-meta [:price :cent-amount-daily]))
                                       (not= (:currency price)
                                             (get-in module-meta [:price :currency]))))]
-          (edit-impl
-            (assoc request
-              :body
-              (cond-> module-meta
-                      price-changed? (-> (assoc :price (merge price (:price module-meta)))
-                                         (set-price (auth/current-active-claim request)))
-                      versions (assoc :versions versions)
-                      compatibility (assoc :compatibility compatibility)))))))
+          (-> request
+              (update-in [:cimi-params :select] disj "compatibility")
+              (assoc :body
+                     (cond-> module-meta
+                             price-changed? (-> (assoc :price (merge price (:price module-meta)))
+                                                (utils/set-price (auth/current-active-claim request)))
+                             versions (assoc :versions versions)))
+              edit-impl))))
     (catch Exception e
       (or (ex-data e) (throw e)))))