+++ code/src/sixsq/nuvla/server/resources/module/utils.clj
@@ -1,9 +1,9 @@
 (ns sixsq.nuvla.server.resources.module.utils
   (:require
-    [clj-yaml.core :as yaml]
     [clojure.edn :as edn]
-    [clojure.set :as set]
     [clojure.string :as str]
+    [sixsq.nuvla.db.filter.parser :as parser]
+    [sixsq.nuvla.pricing.impl :as pricing-impl]
     [sixsq.nuvla.auth.acl-resource :as a]
     [sixsq.nuvla.auth.utils :as auth]
     [sixsq.nuvla.server.resources.common.crud :as crud]
@@ -84,61 +84,6 @@
         (recur (dec i))))))
 
 
-(def ^:const compose-specific-keys
-  #{"devices" "build" "cap_add" "cap_drop" "cgroup_parent" "container_name" "depends_on"
-    "external_links" "network_mode" "restart" "security_opt" "tmpfs" "userns_mode" "privileged"
-    "domainname" "ipc" "mac_address" "shm_size"})
-
-
-(def ^:const swarm-specific-keys #{"configs" "deploy"})
-
-
-(defn docker-compose-services-keys-set
-  [{:strs [services] :as _docker-compose}]
-  (set
-    (when (coll? services)
-      (->> services
-           (map (fn [[_ service-content]]
-                  (when (map? service-content)
-                    (-> service-content keys set))))
-           (reduce set/union)))))
-
-
-(defn list-swarm-unsupported-options
-  [services-keys-set]
-  (set/intersection services-keys-set compose-specific-keys))
-
-
-(defn some-services-has-swarm-options?
-  [services-keys-set]
-  (boolean (some swarm-specific-keys services-keys-set)))
-
-(defn parse-and-throw-when-not-parsable-docker-compose
-  [docker-compose]
-  (try
-    (yaml/parse-string docker-compose :keywords false)
-    (catch Exception _
-      (logu/log-and-throw-400 "Server wasn't able to parse docker-compose yaml!"))))
-
-
-(defn get-compatibility-fields
-  [docker-compose]
-  (let [services-keys-set   (docker-compose-services-keys-set docker-compose)
-        unsupported-options (-> services-keys-set list-swarm-unsupported-options vec)]
-    (if (and (not (some-services-has-swarm-options? services-keys-set))
-             (seq unsupported-options))
-      ["docker-compose" []]
-      ["swarm" unsupported-options])))
-
-
-(defn parse-get-compatibility-fields
-  [subtype docker-compose]
-  (when (is-application? subtype)
-    (-> docker-compose
-        parse-and-throw-when-not-parsable-docker-compose
-        get-compatibility-fields)))
-
-
 (defn retrieve-content-id
   [versions index]
   (let [index (or index (last-index versions))]
@@ -164,6 +109,58 @@
                            (throw (r/ex-not-found
                                     (str "Module version not found: " id)))))]
     (assoc module-meta :content module-content)))
+
+(defn get-vendor-by-query-as-admin
+  [filter-str]
+  (let [options {:cimi-params {:filter (parser/parse-cimi-filter filter-str)}}]
+    (-> (crud/query-as-admin "vendor" options)
+        second
+        first)))
+
+(defn active-claim->account-id
+  [active-claim]
+  (or (some->> active-claim
+               (format "parent='%s'")
+               get-vendor-by-query-as-admin
+               :account-id)
+      (throw (r/ex-response (str "unable to resolve vendor account-id for active-claim '"
+                                 active-claim "' ") 409))))
+
+(defn resolve-vendor-email
+  [{{:keys [account-id]} :price :as module-meta}]
+  (if-let [email (some->> account-id
+                          (format "account-id='%s'")
+                          get-vendor-by-query-as-admin
+                          :email)]
+    (assoc-in module-meta [:price :vendor-email] email)
+    module-meta))
+
+(defn set-price
+  [{{:keys [price-id cent-amount-daily currency follow-customer-trial] :as price}
+    :price name :name path :path :as body}
+   active-claim]
+  (if price
+    (let [product-id (some-> price-id pricing-impl/retrieve-price
+                             pricing-impl/price->map :product-id)
+          account-id (active-claim->account-id active-claim)
+          s-price    (pricing-impl/create-price
+                       (cond-> {"currency"    currency
+                                "unit_amount" cent-amount-daily
+                                "recurring"   {"interval"        "month"
+                                               "aggregate_usage" "sum"
+                                               "usage_type"      "metered"}}
+                               product-id (assoc "product" product-id)
+                               (nil? product-id) (assoc "product_data" {"name"       (or name path)
+                                                                        "unit_label" "day"})))
+          price      (cond-> {:price-id          (pricing-impl/get-id s-price)
+                              :product-id        (pricing-impl/get-product s-price)
+                              :account-id        account-id
+                              :cent-amount-daily cent-amount-daily
+                              :currency          currency}
+                             (some? follow-customer-trial) (assoc :follow-customer-trial follow-customer-trial))]
+      (assoc body :price price))
+    body))
+
 
 (defn can-deploy?
   [{:keys [subtype] :as resource} request]