+++ CHANGELOG.md
@@ -4,7 +4,16 @@
 
 ### Changed
 
+- Module - Let user set compatibility flag
+- Module application - Deprecate unsupported-options
+- ES binding - Support bulk-edit
+- Spec helper - Generic spec validator
+- Spec helper - Request body validator
+- Common body - Spec for common request body specs
+- Nuvlabox - Adds bulk editing endpoints for tags
+- Module - Resolve vendor email
 - Module - New subtype applications set
+- Deployment set - v2
 
 ## [6.0.11] - 2023-02-22